pyproject.toml
[project]
name = "dashvolcano"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bokeh>=3.7.3",
    "geopandas>=1.0.1",
    "geoviews>=1.14.0",
    "hvplot>=0.11.2",
    "openpyxl>=3.1.5",
    "panel>=1.7.1",
    "plotly>=6.0.0",
    "watchfiles>=1.0.4",
    "pymongo>=4.0.0",
    "python-dotenv>=1.0.0",
    "dnspython>=2.0.0",
    "holoviews>=1.17.0",
    "scipy>=1.10.0",
    "param>=1.12.0",
    "cftime>=0.1.0",
]
